~ | Using unsigned long long in main bcuz 64 bits x)

+++ Memory/main.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 #include <stdint.h>
 
-long long int 	Types();
-int 			Displacement();
-int				RetWords();
+long long int 		Types();
+int 				Displacement();
+unsigned long long	RetWords();
 
 int main()
 {
 	printf("%lld\n", Types());
 	printf("%d\n", Displacement());
-	printf("%d\n", RetWords());
+	printf("%llu\n", RetWords());
 }
